feat: add order type, refactored seed to seedOrders
